[design] confirm 모달창 UI
